# docs/openapi_redmine_wiki_approval.yml
openapi: 3.0.0
info:
  title: Redmine Wiki Approval API
  description: >
    API for managing multi-step approval workflows for Redmine Wiki pages. 
    Allows starting approvals, granting/rejecting steps, and tracking history.
  version: 1.0.0
  contact:
      name: FloWalchs
      url: https://github.com/FloWalchs/redmine_wiki_approval
  license:
      name: MIT License
      url: https://opensource.org/licenses/MIT
servers:
  - url: "{protocol}://{domain}"
    description: Your Redmine instance
    variables:
      protocol:
        enum:
          - https
          - http
        default: https
      domain:
        default: 192.168.178.99:3000
        description: Redmine Development Server
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-Redmine-API-Key
      description: Your personal Redmine API key found in your account settings.
    ApiKeyQuery:
      type: apiKey
      in: query
      name: key
      description: API key passed as a URL parameter.
  schemas:
    WikiApprovalStatusResponse:
      allOf:
        - $ref: '#/components/schemas/WikiApprovalWorkflow'
        - type: object
          properties:
            wiki_page:
              $ref: '#/components/schemas/WikiPage'
...
    WikiPage:
      type: object
      properties:
        id:
          type: integer
          example: 30
        wiki_id:
          type: integer
          example: 1
        title:
          type: string
          example: "NewWikiTest"
        created_on:
          type: string
          format: date-time
          example: "2026-03-23T20:32:16Z"
        protected:
          type: boolean
          example: false
        parent_id:
          type: integer
          nullable: true
          example: null

    # Single step in the workflow
    WikiApprovalWorkflowStep:
      type: object
      properties:
        id:
          type: integer
        step:
          type: integer
          description: "Sequence number of the step"
        step_type:
          type: string
          enum: [or, and]
        step_status:
          type: string
          enum: [pending, approved, canceled, rejected]
        principal:
          $ref: '#/components/schemas/Principal'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          
    # The main object
    WikiApprovalWorkflow:
      type: object
      properties:
        id:
          type: integer
        version:
          type: integer
        revision:
          type: integer
        author:
          type: object
          properties:
            id:
              type: integer
            name: 
              type: string
        status:
          type: string
          enum: [canceled, draft, pending, rejected, published, released]
        note:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        wiki_approval_workflow_steps:
          type: array
          items:
            $ref: '#/components/schemas/WikiApprovalWorkflowStep'
...
    Principal:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [User, Group]
      example:
        id: 11
        name: "user four"
        type: "User"
...
security:
  - basicAuth: []
  - ApiKeyHeader: []
  - ApiKeyQuery: []
...
paths:
...
  /projects/{project_id}/wiki_approval/{title}/grant.json:
    put:
      summary: "Submit Approval Decision"
      description: >
        Submits a 'grant' or 'reject' decision for the current approval step. 
        The system automatically identifies the relevant step based on the 
        authenticated user's credentials. If the user is not assigned to the 
        active step, the request will return an error. permission required: wiki_approval_grant
      operationId: wiki_approval_grant
      tags:
        - "Wiki Approval"
      parameters: *wiki_approval
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                step_status:
                  type: string
                  enum: [granted, rejected]
                  default: granted
                  description: "The decision for the current step. 'granted' allows the process to continue, 'rejected' stops it."
                note:
                  type: string
                  description: "An optional comment. **Required** when step_status is 'rejected'."
                  example: "Missing technical documentation in section 3."
            # Using our previously defined example anchor or a new one
            example:
              step_status: "rejected"
              note: "Rejected due to missing security clearance details."
      responses:
        '200':
          description: "Decision recorded successfully. Returns the updated workflow status."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WikiApprovalStatusResponse'
              example: *approval_status
        '403':
          description: "Forbidden. The authenticated user is not assigned to the current active step."
        '404':
          description: "Page not found, or parameter status not found"
        '422':
          description: "Unprocessable Entity. The workflow is not in a state where it can be approved (e.g., already released or canceled)."

  # Forward
  /projects/{project_id}/wiki_approval/{title}/forward.json:
    put:
      summary: Forward Approval
      description: > 
        Reassign the current approval step to another user.
        The system automatically identifies the relevant step based on the authenticated user's credentials. If the user is not assigned to the 
        active step, the request will return an error. permission required: wiki_approval_forward
      operationId: wiki_approval_forward
      tags:
        - "Wiki Approval"
      parameters: *wiki_approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - principal_id
                - note
              properties:
                principal_id:
                  type: integer
                  description: "The ID of the new User or Group to whom the step is being forwarded."
                  example: 15
                note:
                  type: string
                  description: "The reason for forwarding the step (e.g., 'Out of office' or 'Technical expertise required')."
                  example: "Forwarding to the technical lead for final review."
      responses:
        '200':
          description: Approval successfully forwarded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WikiApprovalStatusResponse'
              example: *approval_status
              
        '403':
          description: "Forbidden. The authenticated user is not assigned to the current active step."
        '404':
          description: "Page not found, or principal_id not found"
        '422':
          description: "Unprocessable Entity. The workflow is not in a state where it can be approved, or duplicate users, or missing note parameter."

  # History
...